nosql/pyproject.toml
[tool.poetry]
name = "nosql"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
quickfs = "^0.1.1"
requests = "^2.32.3"
pymongo = "^4.11.1"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
streamlit = "^1.42.2"
plotly = "^6.0.0"
matplotlib = "^3.10.1"
fpdf = "^1.7.2"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
sphinx = "^8.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
